Updated to allow multiple class registrations
Fixing a bug where a DOM element may be an instance of many classes. In
this case, it only needs to be upgraded once, but for each class type.
For example a WSK button may be an instance of both a Button class but
also a Ripple class. Thus, the attribute data-upgraded now contains
classes it has been upgraded to, to allow multiple registrations.