Added ability to listen to upgrades in ComponentHandler
Added extra function to register callback functions on a component class
basis such that you can be alerted when a DOM element gets upgraded.
This is useful in a situation such as this:

1) User decides to use wskCheckbox.
2) When the checkbox is clicked the user wishes to update some parent
node's class (to change the visual style of a table row for example).
3) All they need to do now is register an event listener using
componentHandler.registerUpgradedCallback(JSClass, callback) so that
their callback is called every time any DOM element is upgraded to type
JSClass. The callback function takes 1 parameter which is the
HTMLElement that was upgraded.
4) This saves the user much time and efficiency as they don't need to
"Monitor the DOM" for changes. It also allows them to be notified when
things are added after page load too.